requirements.txt

pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0

transformers>=4.42.0
torch>=2.2.0
sentencepiece>=0.1.99
sentence-transformers>=3.0.0

streamlit>=1.33.0
gradio>=4.36.0

tqdm>=4.66.0
pyyaml>=6.0.1
pydantic>=2.7.0
